Added generic preconditioner option for CG
Added accurate_sum function that can improve the accuracy of inner products.
Switched off right now, but probably we should have a switch for it.
Improves consistency between running on 1 and 2 processors.

Added the Gauss-Seidel preconditioning that can be used by setting Alf=+1
(which is PrecondParam = -1.0 in the #PRECONDITIONER command)