FEAT: New postproc script to put .out files in .tar
Using a tarball is more useful due to the ubiquity of the tar utility
in unix-like systems. This will help with the hassle of seeking in a
concatenated file like what you get from the -m/-M option of
`PostProc.pl`.

TODO: Also compress files with xz which is a more modern utility that
is ubiquotuous as well and has a better compression rate than gzip.